Get page text using a separate thread Avoids test timeouts on complex pages
SVN r27324 |2013-07-02 20:30:58 +0000

+++ src/org/labkey/test/BaseSeleniumWebTest.java
@@ -5493,9 +5493,9 @@ public abstract class BaseSeleniumWebTest implements Cleanable, WebTest
         start = System.currentTimeMillis();
     }
 
-    protected int elapsedSeconds()
+    protected long elapsedSeconds()
     {
-        return (int)((System.currentTimeMillis() - start) / 1000);
+        return (System.currentTimeMillis() - start) / 1000;
     }
 
     /**
